Fix the CI on Appveyor
On windows, composer does not use a symlink when installing binaries but
creates a proxy script instead. This proxy is not a PHP script, so running it
through PHP does not work.
The goal of composer is to allow running the script in a way similar to the
Linux usage relying on shebangs.