Remove the dependency on composer/package-versions-deprecated
This package was added in the past to force keeping support for composer
1 when using doctrine/orm. In the meantime, doctrine/orm has switched to
using that package directly, making this root requirement useless.
And in the next releases of doctrine/orm and doctrine/dbal, they are
migrating to use composer-runtime-api (dropping support for composer 1)
instead of using this plugin, as part of the effort to fade out that
compat plugin.